+++ tools/run-velocity.js
@@ -24,6 +24,8 @@ var runVelocity = function (url) {
 
   // All running browser processes that visit the mirror pages
   var browserProcesses = [];
+  // Maps mirror id to url
+  var mirrorUrls = {};
   var ddpConnection = DDP.connect(url);
 
   var killBrowserProcesses = function () {
@@ -129,13 +131,50 @@ var runVelocity = function (url) {
         }, onReady: function () {
           this.connection.registerStore("velocityMirrors", {
             update: function (msg) {
-              if (msg.msg === "added") {
-                visitWithPhantom(msg.fields.rootUrl);
+              if (isMirrorUrlMessage(msg)) {
+                mirrorUrls[msg.id] = generateMirrorUrl(
+                  msg.fields.rootUrl,
+                  msg.fields.rootUrlPath
+                );
+              }
+              if (isMirrorReadyMessage(msg)) {
+                var mirrorUrl = mirrorUrls[msg.id];
+                if (mirrorUrl) {
+                  visitWithPhantom(mirrorUrl);
+                } else {
+                  Console.error(
+                    "Could not find URL of mirror " +
+                    "with the MongoDB ID " + msg.id
+                  );
+                }
               }
             }
           });
         }
       });
+
+      function isUpdateMessage(msg) {
+        return msg.msg === "added" || msg.msg === "changed";
+      }
+
+      function isMirrorUrlMessage(msg) {
+        return isUpdateMessage(msg) && msg.fields.rootUrl;
+      }
+
+      function isMirrorReadyMessage(msg) {
+        return isUpdateMessage(msg) && msg.fields.state === "ready";
+      }
+
+      function generateMirrorUrl(rootUrl, rootUrlPath) {
+        var mirrorUrl = rootUrl;
+        // Handle the breaking change in velocity:core 0.5
+        // See: https://github.com/meteor-velocity/velocity/issues/260
+        if (rootUrlPath && mirrorUrl.indexOf(rootUrlPath) === -1) {
+          mirrorUrl += rootUrlPath;
+        }
+
+        return mirrorUrl;
+      }
     }
   }, 2000);
 };